Rename native -> oat where possible. We likely want to move the repository to rename what's left.
