+++ .github/workflows/release-notes.yml
@@ -32,6 +32,6 @@ jobs:
         GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
       with:
         tag_name: axon-spring-native-${{ env.VERSION }}
-        release_name:  Axon Spring Native Extension v${{ env.VERSION }}
+        release_name: Axon Spring AOT Extension v${{ env.VERSION }}
         body_path: ${{ env.RELEASE_NOTES_FILE }}
         draft: true
