+++ pom.xml
@@ -18,15 +18,15 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
 
-    <groupId>org.axonframework.extensions.springnative</groupId>
-    <artifactId>axon-spring-native</artifactId>
+    <groupId>org.axonframework.extensions.spring-aot</groupId>
+    <artifactId>axon-spring-aot</artifactId>
     <version>4.8.0-SNAPSHOT</version>
 
     <modelVersion>4.0.0</modelVersion>
 
-    <name>Axon Framework - Spring Native Extension</name>
+    <name>Axon Framework - Spring Ahead of Time</name>
     <description>
-        Extension for Axon Framework providing support for Spring Native.
+        Extension for Axon Framework providing support for Spring's ahead of time optimizations.
     </description>
 
     <inceptionYear>2022</inceptionYear>
