+++ src/main/java/org/axonframework/springboot/aot/AxonRuntimeHints.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.axonframework.springboot.nativex;
+package org.axonframework.springboot.aot;
 
 import org.axonframework.eventhandling.GapAwareTrackingToken;
 import org.axonframework.eventhandling.GlobalSequenceTrackingToken;
