+++ src/main/java/org/axonframework/springboot/aot/MessageHandlerRuntimeHintsRegistrar.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.axonframework.springboot.nativex;
+package org.axonframework.springboot.aot;
 
 import org.axonframework.messaging.Message;
 import org.axonframework.messaging.annotation.AnnotatedHandlerInspector;
