+++ src/test/java/org/axonframework/springboot/aot/MessageHandlerRuntimeHintsRegistrarTest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.axonframework.springboot.nativex;
+package org.axonframework.springboot.aot;
 
 import com.axoniq.someproject.api.ChangeStatusCommand;
 import com.axoniq.someproject.api.SingleChildCommand;
